computer-science/nlp/_src/langchain-sandbox/pyproject.toml
[tool.poetry]
name = "langchain-sandbox"
version = "0.1.0"
description = ""
authors = ["xhiroga <13391129+xhiroga@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.12"
langchain = "^0.0.129"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
google-search-results = "^2.4.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
